Fix compatibility with legacy secrets
We have secrets in our vaults, where it seems that the encrypted key is bigger than a standard AES256 key. We need to allow for such keys to stay compatible. The OpenSSL documentation says that keys are of fixed size, but it seems we rely on an undocumented/undefined behavior.